[image_view2] Separate main function to another cpp file
