+++ jsk_ros_patch/image_view2/CMakeLists.txt
@@ -16,7 +16,7 @@ include_directories(include ${catkin_INCLUDE_DIRS})
 
 # Image viewers
 find_package(OpenCV REQUIRED)
-add_executable(image_view2 image_view2.cpp)
+add_executable(image_view2 image_view2.cpp image_view2_node.cpp)
 target_link_libraries(image_view2 ${OpenCV_LIBS} ${catkin_LIBRARIES})
 add_dependencies(image_view2 ${PROJECT_NAME}_gencpp)
